pocs/mcp-interactive-tool/requirements.txt
mcp>=1.2.0
uvicorn>=0.23.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
